+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
+- **breaking:** Remove `Reflect::type_descriptor` ([#90])
 - **fixed:** Fully qualify `FromReflect` in generated code ([#107])
 - **added:** Implement `Reflect`, and friends, for `Infallible` ([#111])
 
+[#90]: https://github.com/EmbarkStudios/mirror-mirror/pull/90
 [#107]: https://github.com/EmbarkStudios/mirror-mirror/pull/107
 [#111]: https://github.com/EmbarkStudios/mirror-mirror/pull/111
 
@@ -81,8 +83,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # 0.1.2 (03. February, 2023)
 
 - **added:** Add `impl From<Key> for KeyPath` ([#88])
+- **breaking:** Remove `Reflect::type_descriptor`. Instead capture the type
+  descriptor explicitly with `<T as DescribeType::type_descriptor()` ([#90])
 
 [#88]: https://github.com/EmbarkStudios/mirror-mirror/pull/88
+[#90]: https://github.com/EmbarkStudios/mirror-mirror/pull/90
 
 # 0.1.1 (17. January, 2023)